# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Gacha Game Backend API
  version: 1.0.0
  description: API for managing gacha rolls, user inventories, auctions, transactions, and user authentication.
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /transaction/add_transaction:
    post:
      tags:
        - TRANSACTION
      summary: Add a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  example: "1"
                amount:
                  type: number
                  example: 100.0
                type:
                  type: string
                  example: "credit"
      responses:
        '200':
          description: Transaction added successfully
        '500':
          description: Failed to add transaction
  /transaction/get_user_transactions/{user_id}:
    get:
      tags:
        - TRANSACTION
      summary: Retrieve all transactions for a user
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: string
          example: "1"
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    transaction_id:
                      type: string
                      example: "txn_001"
                    user_id:
                      type: string
                      example: "1"
                    amount:
                      type: number
                      example: 100.0
                    type:
                      type: string
                      example: "credit"
                    timestamp:
                      type: string
                      format: date-time
                      example: "2023-10-01T12:00:00Z"
        '400':
          description: User_id not found
        '404':
          description: No transactions found
  /transaction/get_user_transactions:
    get:
      tags:
        - TRANSACTION
      summary: Retrieve all transactions for current user
      responses:
        '200':
          description: Transactions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_id:
                    type: string
                    example: "txn_001"
                  user_id:
                    type: string
                    example: "1"
                  amount:
                    type: number
                    example: 100.0
                  type:
                    type: string
                    example: "credit"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2023-10-01T12:00:00Z"
        '400':
          description: User_id not found
        '403':
          description: Admin can not get their own transactions
        '404':
          description: No transactions found
  /transaction/get_transaction:
    get:
      tags:
        - TRANSACTION
      summary: Retrieve a specific transaction
      parameters:
        - in: query
          name: transaction_id
          required: true
          schema:
            type: string
          example: "txn_001"
      responses:
        '200':
          description: Transaction retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  transaction_id:
                    type: string
                    example: "txn_001"
                  user_id:
                    type: string
                    example: "1"
                  amount:
                    type: number
                    example: 100.0
                  type:
                    type: string
                    example: "credit"
                  timestamp:
                    type: string
                    format: date-time
                    example: "2023-10-01T12:00:00Z"
        '404':
          description: Transaction not found
  /transaction/all:
    get:
      tags:
        - TRANSACTION
      summary: Retrieve all transactions
      description: Retrieve all transactions from the database.
      parameters:
        - in: query
          name: user_id
          required: false
          schema:
            type: string
          example: "1"
      responses:
        '200':
          description: A JSON array of all transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    transaction_id:
                      type: integer
                      example: 1
                    user_id:
                      type: integer
                      example: 123
                    amount:
                      type: number
                      format: integer
                      example: 100
                    date:
                      type: string
                      format: date-time
                      example: "2023-10-01T12:00:00Z"
        '404':
          description: No transactions found
...